requirements.txt
uagents>=0.22.3
pymongo>=4.6.1
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
uagents-core>=0.2.3